agent/requirements.txt
# Core dependencies for agent
websocket-client>=1.4.0
requests>=2.28.1
psutil>=5.9.0
pywin32>=303
pystray>=0.19.4
pillow>=9.2.0

# Dependencies for building installer
pyinstaller>=5.6.2